Move print statement to after determining which algorithm to use
vpClusterGlobalIdx is set to 0 unless LIZ>50, and so the log file is misleading in that it prints that the code is using the new algorithm even though it takes the old route. Moving the print statement to after the conditional statement determining whether the old or new algorithm is used.